Mini update
Added 3 new commands. /connect, which is basically a copy of /transfer but is planned to be upgraded later on. /setblock to set a specific block at a specific location. (/setblock also has to support an oldBlockHandling argument which would select what to do with the old block at the given location. This wasn't implemented yet, but is planned in the next updates.), lastly using a hacky way /setmaxplayers was implemented, which can set the maximum count of players for the current game session. For it to work the value provided in the command has to be equal to or below the player slots set in server.properties. Since this is a hacky way it is not perfect, so yeah, more of a max player limiter lol. Implemented listener for the custom max player method, and am planning to add ability state listener in the upcoming updates. Basically a lot of work was done :p