PlaySound basic implementation, oldBlockHandling for /setblock
Added an util SoundManager class, used to convert specific sound format to a different sound format, such as id to name. Also is used for holding all sound values in all types, and can be used to check if specific value is a specific type of sound. May be updated in the future to change the already exisiting functions or to add more functions. The list of sounds is a sample, and not the full list which is why many sounds won't function yet, looking forward to PRs that put the leftover sounds in their place. Updated /setblock to have the oldBlockHandling argument function, vanilla setblock guaranteed lmao. Added /playsound command that uses packets and is dependent on the SoundManager. None of this was tested, so it might not work :P, just a warning. The plugin is still a W.I.P.