Added /immutableworld command
Added the immutable world command, and implemented a listener for it. The next biggest priority is transferring all commands to the format of the Commando virion, to make the commands more vanilla-like. This will be done by me soon when I figure how exactly Commando virion works and after I do some tests with it, but contributions for it are still welcome :D