Updates
Added /tag command, idk what it is used for in MC but oh well let it be. Changed all the namespaces to follow poggit's rules. Wrote down more permission nodes in plugin.yml, and changed the version to 1.0.0-alpha1 because this is not yet a full release with all known vanilla commands, and because there may be bugs that are not yet found. This commit is planned to be the first pre-release for poggit, although there might be flaws that I have missed that have to be changed to follow poggit rules.